Moved ros_utility into its own folder
