+++ ada_feeding/ada_feeding/trees/stop_servo_tree.py
@@ -20,7 +20,7 @@ from std_msgs.msg import Header
 from std_srvs.srv import Trigger
 
 # Local imports
-from ada_feeding.behaviors import UpdateTimestamp
+from ada_feeding.behaviors.ros import UpdateTimestamp
 from ada_feeding.helpers import BlackboardKey
 from ada_feeding.idioms import retry_call_ros_service, wait_for_secs
 from ada_feeding.trees import TriggerTree
